+++ include/gui/project/window.hpp
@@ -28,117 +28,65 @@
 
 namespace Toolbox::UI {
 
-    class ProjectViewWindow final : public ImWindow {
+    class NewItemWindow final : public ImWindow {
     public:
-        ProjectViewWindow(const std::string &name);
-        ~ProjectViewWindow() = default;
+        NewItemWindow(const std::string &name);
+        ~NewItemWindow() = default;
 
     protected:
-        void onRenderMenuBar() override;
         void onRenderBody(TimeStep delta_time) override;
 
-        void renderProjectTreeView();
-        void renderProjectFolderView();
-        void renderProjectFolderButton();
-        void renderProjectFileButton();
-
-        bool isViewedAncestor(const ModelIndex &index);
-        void renderFolderTree(const ModelIndex &index);
-        void initFolderAssets(const ModelIndex &index);
+        void renderItemRow();
+        void renderItemDescription();
+        void renderControlPanel();
 
     public:
-        ImGuiWindowFlags flags() const override {
-            return ImWindow::flags() | ImGuiWindowFlags_MenuBar;
-        }
+        ImGuiWindowFlags flags() const override { return ImWindow::flags(); }
 
         const ImGuiWindowClass *windowClass() const override {
-            if (parent() && parent()->windowClass()) {
-                return parent()->windowClass();
-            }
-
             ImGuiWindow *currentWindow           = ImGui::GetCurrentWindow();
             m_window_class.ClassId               = (ImGuiID)getUUID();
             m_window_class.ParentViewportId      = currentWindow->ViewportId;
-            m_window_class.DockingAllowUnclassed = true;
+            m_window_class.DockingAllowUnclassed = false;
             m_window_class.DockingAlwaysTabBar   = false;
             return nullptr;
         }
 
         std::optional<ImVec2> minSize() const override {
             return {
-                {600, 400}
+                {600, 700}
             };
         }
 
-        [[nodiscard]] std::string context() const override { return m_project_root.string(); }
+        [[nodiscard]] std::string context() const override { return ""; }
 
         [[nodiscard]] bool unsaved() const override { return false; }
 
         // Returns the supported file types, empty string is designed for a folder.
         [[nodiscard]] std::vector<std::string> extensions() const override { return {}; }
 
-        [[nodiscard]] bool onLoadData(const std::filesystem::path &path) override;
+        [[nodiscard]] bool onLoadData(const std::filesystem::path &path) override { return false; }
 
         [[nodiscard]] bool onSaveData(std::optional<std::filesystem::path> path) override {
-            return true;
+            return false;
         }
 
         void onAttach() override;
         void onDetach() override;
         void onImGuiUpdate(TimeStep delta_time) override;
-        void onContextMenuEvent(RefPtr<ContextMenuEvent> ev) override;
-        void onDragEvent(RefPtr<DragEvent> ev) override;
-        void onDropEvent(RefPtr<DropEvent> ev) override;
-
-        void buildContextMenu();
-        MimeData buildFolderViewMimeData() const;
-
-        // Selection actions
-        void actionDeleteIndexes(std::vector<ModelIndex> &indices);
-        void actionOpenIndexes(const std::vector<ModelIndex> &indices);
-        void actionRenameIndex(const ModelIndex &index);
-        void actionPasteIntoIndex(const ModelIndex &index, const std::vector<fs_path> &data);
-        void actionCopyIndexes(const std::vector<ModelIndex> &indices);
-
-        void actionSelectIndex(const ModelIndex &view_index, const ModelIndex &child_index,
-                                  bool is_selected);
-        void actionClearRequestExcIndex(const ModelIndex &view_index,
-                                        const ModelIndex &child_index, bool is_left_button);
-
-        bool actionOpenScene(const ModelIndex &index);
-        bool actionOpenPad(const ModelIndex &index);
-
-        bool isPathForScene(const ModelIndex &index) const;
 
     private:
-        fs_path m_project_root;
+        using window_constructor = std::function<ScopePtr<ImWindow>(const std::string &)>;
 
-        // TODO: Have filesystem model.
-        FileSystemModelSortFilterProxy m_tree_proxy;
-        FileSystemModelSortFilterProxy m_view_proxy;
-        RefPtr<FileSystemModel> m_file_system_model;
+        struct ItemInfo {
+            std::string m_name;
+            std::string m_extension;
+            std::string m_description;
+            RefPtr<const ImageHandle> m_icon;
+            window_constructor m_win_factory;
+        };
 
-        ModelIndex m_last_selected_index;
-        std::vector<ModelIndex> m_selected_indices;
-        std::vector<ProjectAsset> m_view_assets;
-        ModelIndex m_view_index;
-
-        std::unordered_map<std::string, ImageHandle> m_icon_map;
-        ImagePainter m_icon_painter;
-
-        ContextMenu<ModelIndex> m_folder_view_context_menu;
-        ContextMenu<ModelIndex> m_tree_view_context_menu;
-        std::vector<ModelIndex> m_selected_indices_ctx;
-
-        bool m_is_renaming = false;
-        char m_rename_buffer[128];
-
-        bool m_delete_without_request = false;
-        bool m_delete_requested       = false;
-
-        bool m_did_drag_drop = false;
-
-        ImVec2 m_last_reg_mouse_pos;
+        std::vector<ItemInfo> m_item_infos;
     };
 
 }  // namespace Toolbox::UI
